el formulario ya calcula y escribe, cuidado con el control de errores

+++ crypto/forms.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from flask_wtf import FlaskForm
-from wtforms import (DateTimeField, IntegerField, SelectField, SubmitField,
+from wtforms import (DateTimeField, IntegerField, SelectField, SubmitField,FloatField,
                     validators)
 from wtforms.validators import DataRequired, NumberRange
 
@@ -9,9 +9,9 @@ from wtforms.validators import DataRequired, NumberRange
 class Formulario(FlaskForm):
     momento=DateTimeField(label="fecha",validators=[DataRequired(message="debe especificarse fecha")],format='%Y-%m-%d %H:%M:%S', default=datetime.today())
     moneda_inicial=SelectField(label="moneda_inicial",validators=[DataRequired(message="debe especificarse")],choices=[('EUR'),('BTC'),('ETH'),('XRP'),('LTC'),('BCH'),('BNB'),('USDT'),('EOS'),('BSV'),('XLM'),('ADA'),('TRX')])
-    moneda_inicial_Q=IntegerField(label="moneda_inicial_Q",validators=[DataRequired(message="debe especificar cantidad"), NumberRange(message="debe ser mayor que 0", min=0)])
+    moneda_inicial_Q=FloatField(label="moneda_inicial_Q",validators=[DataRequired(message="debe especificar cantidad"), NumberRange(message="debe ser mayor que 0", min=0)])
     moneda_final=SelectField(label="moneda_final",choices=[('EUR'),('BTC'),('ETH'),('XRP'),('LTC'),('BCH'),('BNB'),('USDT'),('EOS'),('BSV'),('XLM'),('ADA'),('TRX')])
-    moneda_final_Q=IntegerField(label="moneda_final_Q",default=0,validators=[NumberRange(min=0, message="debe ser mayor que 0")])
-    precio_unitario=IntegerField(label="precio_unitario",default=0,validators=[NumberRange(min=0, message="debe ser mayor que 0")])
+    moneda_final_Q=FloatField(label="moneda_final_Q",default=0,validators=[NumberRange(min=0, message="debe ser mayor que 0")])
+    precio_unitario=FloatField(label="precio_unitario",default=0,validators=[NumberRange(min=0, message="debe ser mayor que 0")])
     calcular=SubmitField('Calcular')
     comprar=SubmitField('Comprar')
